+++ components/case/caseController.js
@@ -10,8 +10,18 @@ const {
 const logger = require('../../utils/logger').logger;
 const urls = require('../../utils/urlSlugs');
 const authenticator = require('../../utils/apiAuth');
+const prometheus = require('../../utils/prometheus').prometheus;
+
+
+const caseArchiveStopProcessing = new prometheus.Gauge({
+  name: 'nabu_case_archive_stop_processing',
+  help: 'The case was set to stop processing',
+  labelNames: ['caseIdentifier'],
+})
 
 function arraysEquals (array1, array2) {
+  array1 = array1 || [];
+  array2 = array2 || [];
   return (
     array1.every((item) => array2.includes(item)) &&
     array2.every((item) => array1.includes(item))
@@ -30,7 +40,8 @@ const getCaseArchive = async (req, res, next) => {
       res.status(404).end();
     }
   } catch (e) {
-    handleErrors(e, 'Error getting case', logger, next);
+    logger.error(e);
+    handleErrors(e, `Error getting case ${req.params.caseIdentifier}`, logger, next);
   }
 };
 
@@ -74,6 +85,7 @@ const allCaseArchives = async (req, res, next) => {
     }
     next();
   } catch (e) {
+    logger.error(e);
     handleErrors(e, 'Error getting cases', logger, next);
   }
 };
@@ -87,13 +99,8 @@ const isCompletelyArchived = (kase) => {
   );
 };
 
-const isNotArchived = (kase) => {
-  return (
-    kase.filesCopiedToOffsiteArchiveStagingDir == null &&
-    kase.commvaultBackupJobId == null &&
-    kase.filesLoadedIntoVidarrArchival == null &&
-    kase.caseFilesUnloaded == null
-  );
+const hasArchivingStarted = (kase) => {
+    return kase.filesCopiedToOffsiteArchiveStagingDir != null;
 };
 
 const addCaseArchive = async (req, res, next) => {
@@ -109,49 +116,72 @@ const addCaseArchive = async (req, res, next) => {
       res.status(201).end();
     } else {
       for (let existingCase of existingCases) {
-        if (
-          existingCase.requisitionId == req.body.requisitionId &&
-          arraysEquals(existingCase.limsIds, req.body.limsIds) &&
-          arraysEquals(
+        // check for conflicting changes
+        let errors = [];
+        let caseStoppedProcessing = false;
+        if (existingCase.requisitionId != req.body.requisitionId) {
+          errors.push(`Requisition (${req.body.requisitionId}) from request does not match requisition ${existingCase.requisitionId} for case ${req.body.caseIdentifier}`);
+        }
+        if (!arraysEquals(existingCase.limsIds, req.body.limsIds)) {
+          errors.push(`LIMS IDs (${req.body.limsIds}) from request do not match LIMS IDs ${existingCase.limsIds} for case ${req.body.caseIdentifier}`);
+        }
+        if (hasArchivingStarted(existingCase)) {
+          if (!arraysEquals(
             existingCase.workflowRunIdsForOffsiteArchive,
             req.body.workflowRunIdsForOffsiteArchive
-          ) &&
-          arraysEquals(
+          )) {
+            errors.push(`Requested offsite archive files list ${req.body.workflowRunIdsForOffsiteArchive} does not match offsite files archive list ${existingCase.workflowRunIdsForOffsiteArchive} for case ${existingCase.caseIdentifier}`);
+            await caseDao.setCaseArchiveDoNotProcess(existingCase.caseIdentifier);
+            caseStoppedProcessing = true;
+          }
+          if (!arraysEquals(
             existingCase.workflowRunIdsForVidarrArchival,
             req.body.workflowRunIdsForVidarrArchival
-          )
-        ) {
-          // case data is same, no need to update
-          res.status(200).json(existingCase);
-          return true;
-        } else if (
-          existingCase.requisitionId != req.body.requisitionId ||
-          !arraysEquals(existingCase.limsIds, req.body.limsIds)
-        ) {
-          // data has changed
-          throw new ConflictingDataError(
-            `Cannot modify data for case ${existingCase.caseIdentifier}, requisition or lims identifier(s) do not matching existing data`
-          );
-        } else if (isNotArchived(existingCase)) {
-          // no harm in modifying a case that hasn't yet been archived
-          await upsert(req.body, false);
-          res.status(201).end();
-          return true;
-        } else if (isCompletelyArchived(existingCase)) {
-          // cannot modify a record that has been archived already
-          throw new ConflictingDataError(
-            `Cannot modify data for case ${existingCase.caseIdentifier} that's completed archiving`
-          );
-        } else {
-          // case data is different but files have already been copied to archiving directory / archiving may have begun
-          throw new ConflictingDataError(
-            `Cannot modify data for case ${existingCase.caseIdentifier} that's actively being archived`
-          );
+          )) {
+            errors.push(`Requested onsite archive files list ${req.body.workflowRunIdsForVidarrArchival} does not match onsite archive files list ${existingCase.workflowRunIdsForVidarrArchival} for case ${existingCase.caseIdentifier}`);
+            await caseDao.setCaseArchiveDoNotProcess(existingCase.caseIdentifier);
+            caseStoppedProcessing = true;
+          }
+          if (!arraysEquals(existingCase.archiveWith, req.body.archiveWith)) {
+            errors.push(`Requested archive_with=${req.body.archiveWith} from request does not match archive_with=${existingCase.archiveWith} for case ${req.body.caseIdentifier}`);
+            await caseDao.setCaseArchiveDoNotProcess(existingCase.caseIdentifier);
+            caseStoppedProcessing = true;
+          }
+          if (existingCase.archiveTarget != req.body.archiveTarget) {
+            errors.push(`Archive target '${req.body.archiveTarget}' from request does not match archive target '${existingCase.archiveTarget}' for case ${req.body.caseIdentifier}`);
+            await caseDao.setCaseArchiveDoNotProcess(existingCase.caseIdentifier);
+            caseStoppedProcessing = true;
+          }
         }
+        if (caseStoppedProcessing) {
+          caseArchiveStopProcessing.inc({'caseIdentifier': existingCase.caseIdentifier});
+        }
+        if (errors.length) {
+          for (const e of errors) {
+            logger.error(e);
+          }
+          throw new ConflictingDataError(errors.join("\n"));
+        }
+        if (!hasArchivingStarted(existingCase)) {
+          // can modify a case that hasn't been archived if there are no errors
+          await upsert(req.body, false);
+          res.status(200).end();
+          return true;
+        } else {
+          // if case has started archiving, can only modify case metadata
+          await caseDao.updateMetadata(req.params.caseIdentifier, req.body.metadata);
+          res.status(200).end();
+          return true;
+        }
+        /* NOTE: if we keep encountering HALP actions that don't require any further fixing other than
+           hitting the case/<caseIdentifier>/resume-archiving endpoint because the underlying discrepancy
+           gets resolved upstream, we might want to adjust these last two conditions to also clear the
+           do-not-process flag */
       }
     }
   } catch (e) {
-    handleErrors(e, 'Error adding case', logger, next);
+    logger.error(e);
+    handleErrors(e, `Error adding case ${req.body.caseIdentifier}`, logger, next);
   }
 };
 
@@ -170,6 +200,9 @@ const filesCopiedToOffsiteStagingDir = async (req, res, next) => {
         'Must provide an unload file\'s contents in request body'
       );
     }
+    if (!req.params.batchId) {
+      throw new ValidationError('Must provide a batch ID in URL')
+    }
     const updatedCase = await caseDao.updateFilesCopiedToOffsiteStagingDir(
       req.params.caseIdentifier,
       req.params.batchId,
@@ -181,7 +214,8 @@ const filesCopiedToOffsiteStagingDir = async (req, res, next) => {
       res.status(404).end();
     }
   } catch (e) {
-    handleErrors(e, 'Error updating case', logger, next);
+    logger.error(e);
+    handleErrors(e, `Error updating case ${req.params.caseIdentifier}`, logger, next);
   }
 };
 
@@ -202,7 +236,8 @@ const filesLoadedIntoVidarrArchival = async (req, res, next) => {
       res.status(404).end();
     }
   } catch (e) {
-    handleErrors(e, 'Error updating case', logger, next);
+    logger.error(e);
+    handleErrors(e, `Error updating case ${req.params.caseIdentifier}`, logger, next);
   }
 };
 
@@ -218,7 +253,8 @@ const filesSentOffsite = async (req, res, next) => {
       res.status(404).end();
     }
   } catch (e) {
-    handleErrors(e, 'Error updating case', logger, next);
+    logger.error(e);
+    handleErrors(e, `Error updating case ${req.params.caseIdentifier}`, logger, next);
   }
 };
 
@@ -233,9 +269,22 @@ const caseFilesUnloaded = async (req, res, next) => {
       res.status(404).end();
     }
   } catch (e) {
-    handleErrors(e, 'Error updating case', logger, next);
+    logger.error(e);
+    handleErrors(e, `Error updating case ${req.params.caseIdentifier}`, logger, next);
   }
 };
+
+const resumeCaseArchiveProcessing = async (req, res, next) => {
+  try {
+    await caseDao.resumeCaseArchiveProcessing(req.params.caseIdentifier);
+    caseArchiveStopProcessing.set({'caseIdentifier': req.params.caseIdentifier}, 0);
+    const caseArchive = await caseDao.getByCaseIdentifier(req.params.caseIdentifier, false);
+    res.status(200).json(caseArchive);
+  } catch (e) {
+    logger.error(e);
+    handleErrors(e, `Error resuming processing for case ${req.params.caseIdentifier}`, logger, next);
+  }
+}
 
 module.exports = {
   allCaseArchives: allCaseArchives,
@@ -245,4 +294,5 @@ module.exports = {
   filesCopiedToOffsiteStagingDir: filesCopiedToOffsiteStagingDir,
   filesLoadedIntoVidarrArchival: filesLoadedIntoVidarrArchival,
   filesSentOffsite: filesSentOffsite,
+  resumeCaseArchiveProcessing: resumeCaseArchiveProcessing,
 };