+++ utils/controllerUtils.js
@@ -4,26 +4,26 @@ const NotFoundError = require('./dbUtils').NotFoundError;
 
 /** set up custom error if bad params are given */
 class ValidationError extends Error {
-  constructor (message = '', ...args) {
-    super(message, ...args);
+  constructor (message) {
+    super(message);
     this.name = 'ValidationError';
-    this.message = message || '';
+    this.message = message
   }
 }
 
 class ConflictingDataError extends Error {
-  constructor (message = 'Cannot update with new data', ...args) {
-    super(message, ...args);
+  constructor (message) {
+    super(message);
     this.name = 'ConflictingDataError';
-    this.message = message || '';
+    this.message = message;
   }
 }
 
 class AuthenticationError extends Error {
-  constructor (message = 'Cannot update with current authentication', ...args) {
-    super(message, ...args);
+  constructor (message) {
+    super(message);
     this.name = 'AuthenticationError';
-    this.message = message || '';
+    this.message = message;
   }
 }
 
@@ -40,24 +40,24 @@ function generateError (statusCode, errorMessage) {
 function handleErrors (e, defaultMessage, logger, next) {
   /* eslint-disable */
   if (e instanceof ValidationError) {
-    logger.info(e.message);
+    logger.error(e);
     next(generateError(400, e.message));
   } else if (e instanceof NotFoundError) {
     if (process.env.DEBUG == 'true') console.log(e);
-    logger.info(e);
+    logger.error(e);
     next(generateError(404, null));
   } else if (e instanceof ConflictingDataError) {
     logger.error(e);
     next(generateError(409, e.message));
   } else if (e instanceof AuthenticationError) {
-    logger.info(e);
+    logger.error(e);
     next(generateError(401, e.message));
   } else if (e.status) {
-    logger.info(e);
+    logger.error(e);
     return next(e); // generateError has already been called, usually because it's a user error
   } else {
     console.log(e);
-    logger.debug(e);
+    console.log(e.stack);
     logger.error({ error: e, method: 'handleErrors' });
     next(generateError(500, defaultMessage || 'Error'));
   }